// client/src/features/projects/schemas/project.schemas.ts
import { z } from "zod";

export const PlottingEnum = z.enum([
    "REGULAR", "PROTOTYPE"
]);

export const ScopeEnum = z.enum([
    "NEW_ASSY", "MODIF_MAJOR", "MODIF_MINOR"
]);

export const PROJECT_TYPES = PlottingEnum.options;
export const PROJECT_SCOPES = ScopeEnum.options;

export const createProjectScheme = z.object({
    assyNumber: z.string().min(2, "Assy Number is too short").max(50),
    customer: z.string().min(2, "Customer is too short").max(50),
    
    totalPo: z.number().min(1, "Total PO must be greater than 0"),
    plotting: PlottingEnum,
    scope: ScopeEnum,  
    breakdownDays: z.number().min(1, "Breakdown Days must be at least 1").max(10),

    orderDate: z.string().refine((date) => !isNaN(Date.parse(date)), { message: "Invalid date" }), 
    etd: z.string().refine((date) => !isNaN(Date.parse(date)), { message: "Invalid date" }),
});

export type CreateProjectFormData = z.infer<typeof createProjectScheme>;


// 1. The Production Stage Enum (Matching your Prisma Schema)
export type ProductionStage = 
  | "PLANNING"
  | "PP"
  | "HOUSING"
  | "ASSEMBLING"
  | "JS_ACC"
  | "VISUAL"
  | "CHECKER"
  | "JS_FINISHING"
  | "PREDEL";

// 2. The Entity (What the Database returns)
export interface ProjectEntity {
  id: string;
  assyNumber: string;
  customer: string;
  totalPo: number; 
  plotting: "REGULAR" | "PROTOTYPE";
  scope: "NEW_ASSY" | "MODIF_MAJOR" | "MODIF_MINOR";
  breakdownDays: number;
  
  // Dates come from the API as ISO strings
  orderDate: string;
  etd: string;
  
  // Status fields generated by Prisma
  engineeringStatus: string;
  productionStage: ProductionStage; 
  createdAt: string;
}